requirements.txt
langfuse>=2.0.0
mistralai>=1.0.0
python-dotenv>=1.0.0
